refactor(board): extract useBoardPointerControls (D2 hook extraction, PR 3 of 3)
D2-board-tsx-decomposition-scoping.md Tier-1, final hook. Wheel-zoom,
drag-pan, double-click-to-fit, and the zoom-tray's applyZoomStep /
resetCameraToFit move verbatim into
client/src/components/board/useBoardPointerControls.ts, alongside the
isDragging / dragStart state they own.

Genuinely last of the three, as corrected by PR 2 (#178) — its params
(camera, setCamera, minCameraScale, manualCameraRef, markManualCamera,
fitCameraToContainer(Ref)) are useBoardCamera's outputs.

Board.tsx: 479 -> 396 lines. D2's Tier-1 hook decomposition is complete:
1266 lines (pre-#173) -> 396, split across boardLayout.ts (pure geometry,
#173), useBoardRenderLayout.ts (#177), useBoardCamera.ts (#178), and this
file — each independently testable, the component itself now composing
three hooks + render.

Verification:
- tsc -b clean; full client vitest 230/230 files, 1726/1726 tests
- lint: 48/50 warnings (unchanged); lint:hooks (--max-warnings 0) clean
- board-camera.spec.ts (#176) loop-tested 10/10 clean against the
  complete 3-hook decomposition (render-layout + camera + pointer
  controls all extracted) — the pan/zoom/double-click paths this hook
  owns are exactly what the spec asserts
